requirements.txt
streamlit
pdf2image
pillow
reportlab
